\name{cln.contour}
\alias{cln.contour}
\title{
Contour plot of the conditional logistic normal distribution in \eqn{S^2}
}

\description{
Contour plot of the conditional logistic normal distribution in \eqn{S^2}.
}

\usage{
cln.contour(m, S, n = 100, y = NULL, cont.line = FALSE)
}

\arguments{
\item{m}{
A value with the concentration parameter.
}
\item{S}{
A vector with the mean vector (3 values) in the simplex.
}
\item{n}{
The number of grid points to consider over which the density is calculated.
}
\item{y}{
This is either NULL (no data) or contains a 3 column matrix with
compositional data.
}
\item{cont.line}{
Do you want the contour lines to appear? If yes, set this TRUE.
}
}

\details{
The user can plot only the contour lines of a zero adjusted Dirichlet distribution with
som given parameters, or can also add the relevant data should he/she wish to.
}

\value{
A ternary diagram with the points and the conditional multivariate normal contour lines.
}
\references{
Tsagris M. and Stewart C. (2018). A Dirichlet regression model for compositional data with zeros.
Lobachevskii Journal of Mathematics, 39(3): 398--412.

Preprint available from https://arxiv.org/pdf/1410.5011.pdf
}

\author{
Michail Tsagris.

R implementation and documentation: Michail Tsagris \email{mtsagris@uoc.gr}.
}

\seealso{
\code{ \link{dcln}
}
}

\examples{
y <- as.matrix(iris[, 1:3])
y <- y / rowSums(y)
ind <- sample(150, 10)
for ( k in ind )  y[k, sample(3, 1)] <- 0
y <- y / rowSums(y)
mod <- cln.mle(y)

cln.contour( m = mod$m, S = mod$S )
}

\keyword{ Dirichlet distribution }
\keyword{ contour plot }
