\name{Compositionalcln-package}
\alias{Compositionalcln-package}
\docType{package}
\title{
Modelling Compositional Data with Zero Values
}

\description{
Modelling Compositional Data with Zero Values.
}

\details{
\tabular{ll}{
Package: \tab Compositionalcln\cr
Type: \tab Package\cr
Version: \tab 1.1 \cr
Date: \tab 2026-10-06 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Alzeley O. and Tsagris M. (2026). Modelling compositional data with structural zero values
https://arxiv.org/pdf/2608.29954

Aitchison J. (1986). The statistical analysis of compositional data.
}





