\name{Compositionalcln-package}
\alias{Compositionalcln-package}
\docType{package}
\title{
Modelling Compositional Data With Zero Values
}

\description{
Modelling Compositional Data With Zero Values.
}

\details{
\tabular{ll}{
Package: \tab Compositionalcln\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2026-08-20 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Aitchison J. (1986). The statistical analysis of compositional data.
}





